Feature/8/api overhaul

+++ src/highlightwindow_win32.c
@@ -3,7 +3,6 @@
 #endif 
 
 #include <windows.h>
-#include <stdio.h>
 #include "highlightwindow.h"
 
 #define ID_CLOSE_TIMER 1001
@@ -38,7 +37,7 @@ void showHighlightWindow(int32_t x, int32_t y, int32_t width, int32_t height, lo
         NULL
         );
 
-    SetLayeredWindowAttributes(hwnd, 0, 255 * opacity, LWA_ALPHA);
+    SetLayeredWindowAttributes(hwnd, 0, (BYTE)(255 * opacity), LWA_ALPHA);
 
     if (hwnd == NULL) {
         return;
